tests: separated queries and subscriptions in different folders; increased test timeout
