+++ indexer/jest.config.ts
@@ -8,6 +8,7 @@ export default {
       },
     ],
   },
+  testTimeout: 10000,
   extensionsToTreatAsEsm: ['.ts'],
   moduleNameMapper: {
     '^(\\.{1,2}/.*)\\.js$': '$1',
