+++ indexer/package.json
@@ -76,6 +76,8 @@
     "migrate:down": "dotenv -e .env npx sequelize-cli db:migrate:undo",
     "test:unit": "jest tests/unit/*.test.ts",
     "test:integration": "dotenv -e .env.testing jest tests/integration/*.test.ts",
+    "test:queries": "dotenv -e .env.testing jest tests/integration/queries/*.test.ts",
+    "test:subscriptions": "dotenv -e .env.testing jest tests/integration/subscriptions/*.test.ts",
     "test:file": "dotenv -e .env.testing jest",
     "test:smoke": "yarn compose:up && sleep 5 && jest tests/docker/smoke.test.ts && yarn compose:down",
     "compose:up": "docker-compose -f docker-compose.yml up -d",
