+++ indexer/tests/integration/queries/fungible-chain-accounts-by-public-key.query.test.ts
@@ -1,20 +1,26 @@
 import { GraphQLClient } from 'graphql-request';
-import { fungibleChainAccountsFixture001 } from './fixtures/fungible-chain-accounts/fungible-chain-accounts.fixture.001';
-import { getFungibleChainAccountsQuery } from './builders/fungible-chain-accounts.builder';
+import { fungibleChainAccountsByPublicKeyFixture001 } from '../fixtures/fungible-chain-accounts-by-public-key/fungible-chain-accounts-by-public-key.fixture.001';
+import { getFungibleChainAccountsByPublicKeyQuery } from '../builders/fungible-chain-accounts-by-public-key.builder';
 
 const client = new GraphQLClient(process.env.API_URL ?? 'http://localhost:3001/graphql');
 
-describe('Fungible Chain Accounts', () => {
+describe('Fungible Chain Accounts By Public Key', () => {
   it('#001', async () => {
-    const query = getFungibleChainAccountsQuery({
-      accountName: '20d27f46670fba5aba26a6291a350fc72b46729af1d9776661479c4860c215da',
-      chainIds: ['0'],
+    const query = getFungibleChainAccountsByPublicKeyQuery({
+      publicKey: '20d27f46670fba5aba26a6291a350fc72b46729af1d9776661479c4860c215da',
+      chainId: '0',
     });
 
     const data = await client.request(query);
 
-    data.fungibleChainAccounts.forEach((chainAccount: any, index: number) => {
-      const fixtureChainAccount = fungibleChainAccountsFixture001.data.fungibleChainAccounts[index];
+    // Compare static properties of fungibleChainAccounts
+    expect(data.fungibleChainAccountsByPublicKey).toHaveLength(
+      fungibleChainAccountsByPublicKeyFixture001.data.fungibleChainAccountsByPublicKey.length,
+    );
+
+    data.fungibleChainAccountsByPublicKey.forEach((chainAccount: any, index: number) => {
+      const fixtureChainAccount =
+        fungibleChainAccountsByPublicKeyFixture001.data.fungibleChainAccountsByPublicKey[index];
 
       // Compare basic properties
       expect(chainAccount.accountName).toBe(fixtureChainAccount.accountName);
