+++ indexer/tests/integration/queries/graph-configuration.test.ts
@@ -1,6 +1,6 @@
 import { GraphQLClient } from 'graphql-request';
-import { getGraphConfigurationQuery } from './builders/graph-configuration.builder';
-import { graphConfigurationFixture001 } from './fixtures/graph-configuration/graph-configuration.fixture.001';
+import { getGraphConfigurationQuery } from '../builders/graph-configuration.builder';
+import { graphConfigurationFixture001 } from '../fixtures/graph-configuration/graph-configuration.fixture.001';
 
 const client = new GraphQLClient(process.env.API_URL ?? 'http://localhost:3001/graphql');
