+++ indexer/tests/integration/queries/network-info.query.test.ts
@@ -1,6 +1,6 @@
 import { GraphQLClient } from 'graphql-request';
-import { getNetworkInfoQuery } from './builders/network-info.builder';
-import { networkInfoFixture001 } from './fixtures/network-info/network-info.fixture.001';
+import { getNetworkInfoQuery } from '../builders/network-info.builder';
+import { networkInfoFixture001 } from '../fixtures/network-info/network-info.fixture.001';
 
 const client = new GraphQLClient(process.env.API_URL ?? 'http://localhost:3001/graphql');
